Moved metadata functions to new module, metadata.py
