+++ pilot/control/data.py
@@ -411,12 +411,15 @@ def copytool_in(queues, traces, args):
                 queues.finished_data_in.put(job)
 
                 # now create input file metadata if required by the payload
-                pilot_user = os.environ.get('PILOT_USER', 'generic').lower()
-                user = __import__('pilot.user.%s.setup' % pilot_user, globals(), locals(), [pilot_user], -1)
-                file_dictionary = get_input_file_dictionary(job.indata)
-                log.debug('file_dictionary=%s' % str(file_dictionary))
-                xml = user.create_input_file_metadata(file_dictionary, job.workdir)
-                log.info('created input file metadata:\n%s' % xml)
+                try:
+                    pilot_user = os.environ.get('PILOT_USER', 'generic').lower()
+                    user = __import__('pilot.user.%s.metadata' % pilot_user, globals(), locals(), [pilot_user], -1)
+                    file_dictionary = get_input_file_dictionary(job.indata)
+                    log.debug('file_dictionary=%s' % str(file_dictionary))
+                    xml = user.create_input_file_metadata(file_dictionary, job.workdir)
+                    log.info('created input file metadata:\n%s' % xml)
+                except Exception as e:
+                    pass
             else:
                 log.warning('stage-in failed, adding job object to failed_data_in queue')
                 job.piloterrorcodes, job.piloterrordiags = errors.add_error_code(errors.STAGEINFAILED)
